Renamed packages back to seedu.duke
