+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,10 +1,10 @@
-package seedu.commoncents;
+package seedu.duke;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class CommonCentsTest {
+class DukeTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
